--- file_processing/tools/__init__.py
@@ -1,3 +1,0 @@
-# from .file_processor_strategy import FileProcessorStrategy
-# from .ocr_decorator import OCRDecorator
-# from .transcription_decorator import TranscriptionDecorator
